Struct fc_cli::FrontierDbCmd

pub struct FrontierDbCmd {
    pub operation: Operation,
    pub column: Column,
    pub key: String,
    pub value: Option<PathBuf>,
    pub shared_params: SharedParams,
    pub pruning_params: PruningParams,
}

Cli tool to interact with the Frontier backend db

Fields
operation: Operation

Specify the operation to perform.

Can be one of create | read | update | delete.

column: Column

Specify the column to query.

Can be one of meta | block | transaction.

key: String

Specify the key to either read or write.

value: Option<PathBuf>

Specify the value to write.

When Some, path to file.

When None, read from stdin.

In any case, payload must be serializable to a known type.

shared_params: SharedParams

Shared parameters

pruning_params: PruningParams
Implementations

impl FrontierDbCmd

pub fn run<C, B: BlockT>(
&self,
client: Arc<C>,
backend: Arc<Backend<B>>
) -> Result<()> where
C: ProvideRuntimeApi<B>,
C::Api: EthereumRuntimeRPCApi<B>,
C: HeaderBackend<B>,